+++ src/lib.rs
@@ -5,6 +5,7 @@
 extern crate ndarray;
 extern crate nix;
 extern crate raw_tty;
+extern crate ropey;
 extern crate smallvec;
 extern crate termion;
 extern crate unicode_segmentation;