+++ src/widget/builtin/mod.rs
@@ -3,8 +3,10 @@ pub mod lineedit;
 pub mod logviewer;
 pub mod promptline;
 pub mod table;
+pub mod textedit;
 
 pub use self::lineedit::*;
 pub use self::logviewer::*;
 pub use self::promptline::*;
 pub use self::table::*;
+pub use self::textedit::*;